Apply rounding to basic descriptive statistics.

Not all statistics should be rounded in the same way, or at all. This function will apply rounding selectively to a list of statistics as returned by stats.default. In particular we don't round counts (N, NMISS and FREQ), and for MIN, MAX and MEDIAN the digits is interpreted as the minimum number of significant digits, so that we don't loose any precision. Percentages are rounded to a fixed number of decimal places (default 1) rather than a specific number of significant digits.

Usage

stats.apply.rounding(
  x,
  digits = 3,
  digits.pct = 1,
  round.median.min.max = TRUE,
  round.integers = TRUE,
  round5up = TRUE,
  rounding.fn = signif_pad,
  ...
)

Arguments

x

A list, such as that returned by stats.default.

digits

An integer specifying the number of significant digits to keep.

digits.pct

An integer specifying the number of digits after the decimal place for percentages.

round.median.min.max

Should rounding applied to median, min and max?

round.integers

Should rounding be limited to digits to the right of the decimal point?

round5up

Should numbers with 5 as the last digit always be rounded up? The standard R approach is "go to the even digit" (IEC 60559 standard, see round), while some other softwares (e.g. SAS, Excel) always round up.

rounding.fn

The function to use to do the rounding. Defaults to signif_pad.

...

Further arguments.

Value

A list with the same number of elements as x. The rounded values will be character (not numeric) and will have 0 padding to ensure consistent number of significant digits.

See also

Examples

x <- round(exp(rnorm(100, 1, 1)), 6)
stats.default(x)
#> $N
#> [1] 100
#> 
#> $NMISS
#> [1] 0
#> 
#> $SUM
#> [1] 416.826
#> 
#> $MEAN
#> [1] 4.16826
#> 
#> $SD
#> [1] 4.466171
#> 
#> $CV
#> [1] 107.1471
#> 
#> $GMEAN
#> [1] 2.739768
#> 
#> $GSD
#> [1] 2.52568
#> 
#> $GCV
#> [1] 116.5947
#> 
#> $MEDIAN
#> [1] 2.926633
#> 
#> $MIN
#> [1] 0.281714
#> 
#> $MAX
#> [1] 26.90584
#> 
#> $q01
#>        1% 
#> 0.4105239 
#> 
#> $q02.5
#>      2.5% 
#> 0.5351426 
#> 
#> $q05
#>        5% 
#> 0.6353341 
#> 
#> $q10
#>       10% 
#> 0.6650688 
#> 
#> $q25
#>      25% 
#> 1.408417 
#> 
#> $q50
#>      50% 
#> 2.926633 
#> 
#> $q75
#>      75% 
#> 4.998712 
#> 
#> $q90
#>      90% 
#> 9.251978 
#> 
#> $q95
#>      95% 
#> 12.40016 
#> 
#> $q97.5
#>    97.5% 
#> 16.43128 
#> 
#> $q99
#>      99% 
#> 22.93844 
#> 
#> $Q1
#>      25% 
#> 1.408417 
#> 
#> $Q2
#>      50% 
#> 2.926633 
#> 
#> $Q3
#>      75% 
#> 4.998712 
#> 
#> $IQR
#>      75% 
#> 3.590295 
#> 
#> $T1
#> 33.33333% 
#>  1.828447 
#> 
#> $T2
#> 66.66667% 
#>  4.124893 
#> 
stats.apply.rounding(stats.default(x), digits=3)
#> $N
#> [1] "100"
#> 
#> $NMISS
#> [1] "0"
#> 
#> $SUM
#> [1] "417"
#> 
#> $MEAN
#> [1] "4.17"
#> 
#> $SD
#> [1] "4.47"
#> 
#> $CV
#> [1] "107.1"
#> 
#> $GMEAN
#> [1] "2.74"
#> 
#> $GSD
#> [1] "2.53"
#> 
#> $GCV
#> [1] "116.6"
#> 
#> $MEDIAN
#> [1] "2.93"
#> 
#> $MIN
#> [1] "0.282"
#> 
#> $MAX
#> [1] "26.9"
#> 
#> $q01
#>      1% 
#> "0.411" 
#> 
#> $q02.5
#>    2.5% 
#> "0.535" 
#> 
#> $q05
#>      5% 
#> "0.635" 
#> 
#> $q10
#>     10% 
#> "0.665" 
#> 
#> $q25
#>    25% 
#> "1.41" 
#> 
#> $q50
#>    50% 
#> "2.93" 
#> 
#> $q75
#>    75% 
#> "5.00" 
#> 
#> $q90
#>    90% 
#> "9.25" 
#> 
#> $q95
#>    95% 
#> "12.4" 
#> 
#> $q97.5
#>  97.5% 
#> "16.4" 
#> 
#> $q99
#>    99% 
#> "22.9" 
#> 
#> $Q1
#>    25% 
#> "1.41" 
#> 
#> $Q2
#>    50% 
#> "2.93" 
#> 
#> $Q3
#>    75% 
#> "5.00" 
#> 
#> $IQR
#>    75% 
#> "3.59" 
#> 
#> $T1
#> 33.33333% 
#>    "1.83" 
#> 
#> $T2
#> 66.66667% 
#>    "4.12" 
#> 
stats.apply.rounding(stats.default(round(x, 1)), digits=3)
#> $N
#> [1] "100"
#> 
#> $NMISS
#> [1] "0"
#> 
#> $SUM
#> [1] "417"
#> 
#> $MEAN
#> [1] "4.17"
#> 
#> $SD
#> [1] "4.47"
#> 
#> $CV
#> [1] "107.2"
#> 
#> $GMEAN
#> [1] "2.73"
#> 
#> $GSD
#> [1] "2.53"
#> 
#> $GCV
#> [1] "117.0"
#> 
#> $MEDIAN
#> [1] "2.95"
#> 
#> $MIN
#> [1] "0.300"
#> 
#> $MAX
#> [1] "26.9"
#> 
#> $q01
#>      1% 
#> "0.399" 
#> 
#> $q02.5
#>    2.5% 
#> "0.548" 
#> 
#> $q05
#>      5% 
#> "0.600" 
#> 
#> $q10
#>     10% 
#> "0.700" 
#> 
#> $q25
#>    25% 
#> "1.40" 
#> 
#> $q50
#>    50% 
#> "2.95" 
#> 
#> $q75
#>    75% 
#> "5.00" 
#> 
#> $q90
#>    90% 
#> "9.28" 
#> 
#> $q95
#>    95% 
#> "12.4" 
#> 
#> $q97.5
#>  97.5% 
#> "16.4" 
#> 
#> $q99
#>    99% 
#> "22.9" 
#> 
#> $Q1
#>    25% 
#> "1.40" 
#> 
#> $Q2
#>    50% 
#> "2.95" 
#> 
#> $Q3
#>    75% 
#> "5.00" 
#> 
#> $IQR
#>    75% 
#> "3.60" 
#> 
#> $T1
#> 33.33333% 
#>    "1.80" 
#> 
#> $T2
#> 66.66667% 
#>    "4.10" 
#>