Compute some basic descriptive statistics.

Values of type factor, character and logical are treated as categorical. For logicals, the two categories are given the labels `Yes` for TRUE, and `No` for FALSE. Factor levels with zero counts are retained.

Usage

stats.default(x, quantile.type = 7, ...)

Arguments

x

A vector or numeric, factor, character or logical values.

quantile.type

An integer from 1 to 9, passed as the type argument to function quantile.

...

Further arguments (ignored).

Value

A list. For numeric x, the list contains the numeric elements:

N: the number of non-missing values

NMISS: the number of missing values

SUM: the sum of the non-missing values

MEAN: the mean of the non-missing values

SD: the standard deviation of the non-missing values

MIN: the minimum of the non-missing values

MEDIAN: the median of the non-missing values

CV: the percent coefficient of variation of the non-missing values

GMEAN: the geometric mean of the non-missing values if non-negative, or NA

GSD: the geometric standard deviation of the non-missing values if non-negative, or NA

GCV: the percent geometric coefficient of variation of the non-missing values if non-negative, or NA

qXX: various quantiles (percentiles) of the non-missing values (q01: 1%, q02.5: 2.5%, q05: 5%, q10: 10%, q25: 25% (first quartile), q33.3: 33.33333% (first tertile), q50: 50% (median, or second quartile), q66.7: 66.66667% (second tertile), q75: 75% (third quartile), q90: 90%, q95: 95%, q97.5: 97.5%, q99: 99%)

Q1: the first quartile of the non-missing values (alias q25)

Q2: the second quartile of the non-missing values (alias q50 or Median)

Q3: the third quartile of the non-missing values (alias q75)

IQR: the inter-quartile range of the non-missing values (i.e., Q3 - Q1)

T1: the first tertile of the non-missing values (alias q33.3)

T2: the second tertile of the non-missing values (alias q66.7)

If x is categorical (i.e. factor, character or logical), the list contains a sublist for each category, where each sublist contains the numeric elements:

FREQ: the frequency count

PCT: the percent relative frequency, including NA in the denominator

PCTnoNA: the percent relative frequency, excluding NA from the denominator

NMISS: the number of missing values

Examples

x <- exp(rnorm(100, 1, 1))
stats.default(x)
#> $N
#> [1] 100
#> 
#> $NMISS
#> [1] 0
#> 
#> $SUM
#> [1] 521.1011
#> 
#> $MEAN
#> [1] 5.211011
#> 
#> $SD
#> [1] 11.57429
#> 
#> $CV
#> [1] 222.1122
#> 
#> $GMEAN
#> [1] 2.933438
#> 
#> $GSD
#> [1] 2.602334
#> 
#> $GCV
#> [1] 122.314
#> 
#> $MEDIAN
#> [1] 2.851331
#> 
#> $MIN
#> [1] 0.3002564
#> 
#> $MAX
#> [1] 112.8022
#> 
#> $q01
#>        1% 
#> 0.3371718 
#> 
#> $q02.5
#>      2.5% 
#> 0.3953798 
#> 
#> $q05
#>        5% 
#> 0.6640415 
#> 
#> $q10
#>      10% 
#> 1.024823 
#> 
#> $q25
#>      25% 
#> 1.678614 
#> 
#> $q50
#>      50% 
#> 2.851331 
#> 
#> $q75
#>      75% 
#> 5.400385 
#> 
#> $q90
#>      90% 
#> 10.31669 
#> 
#> $q95
#>      95% 
#> 14.40168 
#> 
#> $q97.5
#>    97.5% 
#> 16.32419 
#> 
#> $q99
#>      99% 
#> 23.53208 
#> 
#> $Q1
#>      25% 
#> 1.678614 
#> 
#> $Q2
#>      50% 
#> 2.851331 
#> 
#> $Q3
#>      75% 
#> 5.400385 
#> 
#> $IQR
#>      75% 
#> 3.721772 
#> 
#> $T1
#> 33.33333% 
#>  1.820745 
#> 
#> $T2
#> 66.66667% 
#>  3.918019 
#> 

y <- factor(sample(0:1, 99, replace=TRUE), labels=c("Female", "Male"))
y[1:10] <- NA
stats.default(y)
#> $Female
#> $Female$FREQ
#> [1] 37
#> 
#> $Female$PCT
#> [1] 37.37374
#> 
#> $Female$PCTnoNA
#> [1] 41.57303
#> 
#> $Female$NMISS
#> [1] 10
#> 
#> 
#> $Male
#> $Male$FREQ
#> [1] 52
#> 
#> $Male$PCT
#> [1] 52.52525
#> 
#> $Male$PCTnoNA
#> [1] 58.42697
#> 
#> $Male$NMISS
#> [1] 10
#> 
#> 
stats.default(is.na(y))
#> $Yes
#> $Yes$FREQ
#> [1] 10
#> 
#> $Yes$PCT
#> [1] 10.10101
#> 
#> $Yes$PCTnoNA
#> [1] 10.10101
#> 
#> $Yes$NMISS
#> [1] 0
#> 
#> 
#> $No
#> $No$FREQ
#> [1] 89
#> 
#> $No$PCT
#> [1] 89.89899
#> 
#> $No$PCTnoNA
#> [1] 89.89899
#> 
#> $No$NMISS
#> [1] 0
#> 
#>