Convert to HTML table rows.

Many functions exist in R to generate HTML tables. These functions are useful for generating HTML table fragments (rather than whole tables), which can then be used to build up complete tables. The first column my be used to label the rows of the table. Row labels, if specified, can have a special HTML class designated, which can be useful as a hook to customize their appearance using CSS. The same is true for the the first and last row of cells.

Usage

table.rows(
  x,
  row.labels = rownames(x),
  th = FALSE,
  class = NULL,
  rowlabelclass = "rowlabel",
  firstrowclass = "firstrow",
  lastrowclass = "lastrow",
  ...
)

table.data(
  x,
  row.labels = rownames(x),
  th = FALSE,
  class = NULL,
  rowlabelclass = "rowlabel",
  firstrowclass = "firstrow",
  lastrowclass = "lastrow",
  ...
)

Arguments

x

A vector or table-like structure (e.g. a data.frame or matrix).

row.labels

Values for the first column, typically used to label the row, or NULL to omit.

th

A logical. Should th tags be used rather than td?

class

HTML class attribute. Can be a single character, a vector or a matrix.

rowlabelclass

HTML class attribute for the row labels (i.e. first column).

firstrowclass

HTML class attribute for the first row of cells.

lastrowclass

HTML class attribute for the last row of cells.

...

Additional arguments.

Value

A character which contains an HTML table fragment.

Functions

table.data(): Convert to HTML table data (cells).

Examples

x <- matrix(signif_pad(exp(rnorm(5*5, 1, 1))), 5, 5)
table.data(x)
#> [1] "<td class='firstrow'>3.44</td>\n<td class='firstrow'>4.27</td>\n<td class='firstrow'>0.631</td>\n<td class='firstrow'>2.09</td>\n<td class='firstrow'>0.630</td>"
#> [2] "<td>2.94</td>\n<td>2.83</td>\n<td>5.70</td>\n<td>0.564</td>\n<td>5.41</td>"                                                                                      
#> [3] "<td>1.04</td>\n<td>1.78</td>\n<td>18.3</td>\n<td>0.598</td>\n<td>22.2</td>"                                                                                      
#> [4] "<td>2.53</td>\n<td>0.349</td>\n<td>0.642</td>\n<td>0.548</td>\n<td>0.750</td>"                                                                                   
#> [5] "<td class='lastrow'>11.5</td>\n<td class='lastrow'>8.43</td>\n<td class='lastrow'>5.48</td>\n<td class='lastrow'>1.60</td>\n<td class='lastrow'>5.98</td>"       
cat(table.rows(x, NULL))
#> <tr>
#> <td class='rowlabel firstrow'></td>
#> <td class='firstrow'>3.44</td>
#> <td class='firstrow'>4.27</td>
#> <td class='firstrow'>0.631</td>
#> <td class='firstrow'>2.09</td>
#> <td class='firstrow'>0.630</td>
#> </tr>
#> <tr>
#> <td class='rowlabel'></td>
#> <td>2.94</td>
#> <td>2.83</td>
#> <td>5.70</td>
#> <td>0.564</td>
#> <td>5.41</td>
#> </tr>
#> <tr>
#> <td class='rowlabel'></td>
#> <td>1.04</td>
#> <td>1.78</td>
#> <td>18.3</td>
#> <td>0.598</td>
#> <td>22.2</td>
#> </tr>
#> <tr>
#> <td class='rowlabel'></td>
#> <td>2.53</td>
#> <td>0.349</td>
#> <td>0.642</td>
#> <td>0.548</td>
#> <td>0.750</td>
#> </tr>
#> <tr>
#> <td class='rowlabel lastrow'></td>
#> <td class='lastrow'>11.5</td>
#> <td class='lastrow'>8.43</td>
#> <td class='lastrow'>5.48</td>
#> <td class='lastrow'>1.60</td>
#> <td class='lastrow'>5.98</td>
#> </tr>
cat(table.rows(x, LETTERS[1:nrow(x)]))
#> <tr>
#> <td class='rowlabel firstrow'>A</td>
#> <td class='firstrow'>3.44</td>
#> <td class='firstrow'>4.27</td>
#> <td class='firstrow'>0.631</td>
#> <td class='firstrow'>2.09</td>
#> <td class='firstrow'>0.630</td>
#> </tr>
#> <tr>
#> <td class='rowlabel'>B</td>
#> <td>2.94</td>
#> <td>2.83</td>
#> <td>5.70</td>
#> <td>0.564</td>
#> <td>5.41</td>
#> </tr>
#> <tr>
#> <td class='rowlabel'>C</td>
#> <td>1.04</td>
#> <td>1.78</td>
#> <td>18.3</td>
#> <td>0.598</td>
#> <td>22.2</td>
#> </tr>
#> <tr>
#> <td class='rowlabel'>D</td>
#> <td>2.53</td>
#> <td>0.349</td>
#> <td>0.642</td>
#> <td>0.548</td>
#> <td>0.750</td>
#> </tr>
#> <tr>
#> <td class='rowlabel lastrow'>E</td>
#> <td class='lastrow'>11.5</td>
#> <td class='lastrow'>8.43</td>
#> <td class='lastrow'>5.48</td>
#> <td class='lastrow'>1.60</td>
#> <td class='lastrow'>5.98</td>
#> </tr>
cat(table.rows(LETTERS[1:3], "Headings", th=TRUE))
#> <tr>
#> <th class='rowlabel firstrow lastrow'>Headings</th>
#> <th class='firstrow lastrow'>A</th>
#> <th class='firstrow lastrow'>B</th>
#> <th class='firstrow lastrow'>C</th>
#> </tr>